Make read and readtex instructions work. Add --tex commandline flag
This involved adding instructions and pushing around the
information of whether Tex should be used through the entire
codebase. Getting the instructions to get handled correctly in
Reader.hs was particularly confusing. I did it by explicitly adding
another parameter to to the (very confusing) reader loop.

Now it remains to make the tests work.