Tokenize `\` and `$` away in order to support latex math mode
While this obviously isn't the perfect solution, it is a quick and
transparent solution that covers basic usecases.

Moreover, now `"` is also being tokenized away, because it is
intended to be used for telling latex to use text-mode inside a
math-mode environment. This needs a special tex snippet that
remaps `"`. An example on how to use these latex features is in
`examples/powerset.tex`.

Optimally, one would write a tokenizer specifically for math mode,
that then allows the parser to choose how to assign meaning to
specific latex constructs. E.g. one could let fonts specifications
like `\matcal{C}` be part of the variable identifier and let things
like `\frac{x}{y}` be operators. Moreover, there is much more
semantic information that may be potentially attached to latex
constructs. But it would probably be a waste of time to implement
this 'proper' solution with our current tokenizer(consisting of
low-level Text manipulations). One should wait until one gets a
proper tokenizer in order to do this.